+++ litt/cache/cached_table.go
@@ -7,6 +7,7 @@ import (
 	"github.com/Layr-Labs/eigenda/common/cache"
 	"github.com/Layr-Labs/eigenda/litt"
 	"github.com/Layr-Labs/eigenda/litt/types"
+	"github.com/Layr-Labs/eigenda/litt/util"
 )
 
 var _ litt.ManagedTable = &cachedTable{}
@@ -52,14 +53,15 @@ func (c *cachedTable) PutBatch(batch []*types.KVPair) error {
 		return err
 	}
 	for _, kv := range batch {
-		c.cache.Put(string(kv.Key), kv.Value)
+		c.cache.Put(util.UnsafeBytesToString(kv.Key), kv.Value)
 	}
 	return nil
 }
 
 func (c *cachedTable) Get(key []byte) ([]byte, bool, error) {
+	stringKey := util.UnsafeBytesToString(key)
 
-	value, ok := c.cache.Get(string(key))
+	value, ok := c.cache.Get(stringKey)
 	if ok {
 		return value, true, nil
 	}
@@ -70,14 +72,14 @@ func (c *cachedTable) Get(key []byte) ([]byte, bool, error) {
 	}
 
 	if ok {
-		c.cache.Put(string(key), value)
+		c.cache.Put(stringKey, value)
 	}
 
 	return value, ok, nil
 }
 
 func (c *cachedTable) Exists(key []byte) (bool, error) {
-	_, ok := c.cache.Get(string(key))
+	_, ok := c.cache.Get(util.UnsafeBytesToString(key))
 	if ok {
 		return true, nil
 	}
