+++ litt/disktable/keymap/mem_keymap.go
@@ -5,6 +5,7 @@ import (
 	"sync"
 
 	"github.com/Layr-Labs/eigenda/litt/types"
+	"github.com/Layr-Labs/eigenda/litt/util"
 	"github.com/Layr-Labs/eigensdk-go/logging"
 )
 
@@ -26,7 +27,6 @@ type memKeymap struct {
 
 // NewMemKeymap creates a new in-memory keymap.
 func NewMemKeymap(logger logging.Logger, doubleWriteProtection bool) Keymap {
-
 	return &memKeymap{
 		logger:                logger,
 		data:                  make(map[string]types.Address),
@@ -39,14 +39,16 @@ func (m *memKeymap) Put(pairs []*types.KAPair) error {
 	defer m.lock.Unlock()
 
 	for _, pair := range pairs {
+		stringKey := util.UnsafeBytesToString(pair.Key)
+
 		if m.doubleWriteProtection {
-			_, ok := m.data[string(pair.Key)]
+			_, ok := m.data[stringKey]
 			if ok {
 				return fmt.Errorf("key %s already exists", pair.Key)
 			}
 		}
 
-		m.data[string(pair.Key)] = pair.Address
+		m.data[stringKey] = pair.Address
 	}
 	return nil
 }
@@ -55,7 +57,7 @@ func (m *memKeymap) Get(key []byte) (types.Address, bool, error) {
 	m.lock.RLock()
 	defer m.lock.RUnlock()
 
-	address, ok := m.data[string(key)]
+	address, ok := m.data[util.UnsafeBytesToString(key)]
 	return address, ok, nil
 }
 
@@ -64,7 +66,7 @@ func (m *memKeymap) Delete(keys []*types.KAPair) error {
 	defer m.lock.Unlock()
 
 	for _, key := range keys {
-		delete(m.data, string(key.Key))
+		delete(m.data, util.UnsafeBytesToString(key.Key))
 	}
 
 	return nil
