+++ litt/util/fatal_error_handler.go
@@ -22,33 +22,24 @@ type FatalErrorHandler struct {
 
 	logger logging.Logger
 
-	// callback is called when the DB panics.
-	callback func(error)
-
 	// If this is non-nil, the DB is in a "panic" state and will refuse to do additional work.
 	error atomic.Pointer[error]
 }
 
-// NewFatalErrorHandler creates a new FatalErrorHandler struct. Executes the callback function when/if the DB panics.
-// The callback is ignored if it is nil.
-func NewFatalErrorHandler(
-	ctx context.Context,
-	logger logging.Logger,
-	callback func(error)) *FatalErrorHandler {
-
+// NewFatalErrorHandler creates a new FatalErrorHandler struct.
+func NewFatalErrorHandler(ctx context.Context, logger logging.Logger) *FatalErrorHandler {
 	ctx, cancel := context.WithCancel(ctx)
 
 	return &FatalErrorHandler{
-		ctx:      ctx,
-		cancel:   cancel,
-		logger:   logger,
-		callback: callback,
+		ctx:    ctx,
+		cancel: cancel,
+		logger: logger,
 	}
 }
 
-// AwaitIfNotFatal waits for a value to be sent on a channel. If the channel sends a value, the value is returned.
-// If the DB panics before the channel sends a value, an error is returned.
-func AwaitIfNotFatal[T any](handler *FatalErrorHandler, channel <-chan T) (T, error) {
+// Await waits for a value to be sent on a channel. If the channel sends a value, the value is returned. If the DB
+// panics before the channel sends a value, an error is returned.
+func Await[T any](handler *FatalErrorHandler, channel <-chan T) (T, error) {
 	select {
 	case value := <-channel:
 		return value, nil
@@ -58,9 +49,9 @@ func AwaitIfNotFatal[T any](handler *FatalErrorHandler, channel <-chan T) (T, er
 	}
 }
 
-// SendIfNotFatal sends a value on a channel. If the value is sent, nil is returned. If the DB panics before the value
-// is sent, an error is returned.
-func SendIfNotFatal[T any](handler *FatalErrorHandler, channel chan<- any, value T) error {
+// Send sends a value on a channel. If the value is sent, nil is returned. If the DB panics before the value is sent,
+// an error is returned.
+func Send[T any](handler *FatalErrorHandler, channel chan<- any, value T) error {
 	select {
 	case channel <- value:
 		return nil
@@ -107,14 +98,10 @@ func (h *FatalErrorHandler) Panic(err error) {
 	h.logger.Errorf("LittDB encountered an unrecoverable error: %v\n%s", err, stackTrace)
 
 	// only store the error if there isn't already an error stored
-	firstError := h.error.CompareAndSwap(nil, &err)
+	h.error.CompareAndSwap(nil, &err)
 
 	// Always cancel the context, even if this is not the first error. It's possible that the first "error" was
 	// actually a shutdown request, and we want to make sure that the context is always cancelled in the event
 	// of an unexpected error.
 	h.cancel()
-
-	if firstError && h.callback != nil {
-		h.callback(err)
-	}
 }
